arquitetura/pub-sub/notifier-app/requirements.txt
confluent-kafka==1.5.1
requests
python-telegram-bot
